requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
blessed==1.20.0
boto3==1.34.150
botocore==1.34.162
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.3
crispy-bootstrap4==2024.1
Django==4.2.14
django-crispy-forms==2.3
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.7
jmespath==1.0.1
pathspec==0.10.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.1.9
